+++ Client/Helpers/SwiftUIHelpers.swift
@@ -35,11 +35,11 @@ extension KeyboardReadable {
         Publishers.Merge(
             NotificationCenter.default
                 .publisher(for: UIResponder.keyboardWillShowNotification)
-                .map { ($0.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect)?.height ?? CGFloat(0) },
+                .map { ($0.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect)?.height ?? 0 },
 
             NotificationCenter.default
                 .publisher(for: UIResponder.keyboardWillHideNotification)
-                .map { _ in CGFloat(0) }
+                .map { _ in 0 }
         )
         .eraseToAnyPublisher()
     }
